repository and repository test modified
targetPrice from int to Long and active profile annotation added to the
test

+++ attsd-spring-project/src/main/java/com/examples/spring/project/repositories/DressShopRepository.java
@@ -12,7 +12,7 @@ public interface DressShopRepository extends JpaRepository<DressShop, Long> {
 
 
 	@Query("Select ds from DressShop ds where ds.targetPrice = :n")
-	List<DressShop> findAllDressShopsWithTheSameTargetPrice(@Param("n") int n);
+	List<DressShop> findAllDressShopsWithTheSameTargetPrice(@Param("n") Long n);
 	
 
 	@Query("Select ds from DressShop ds order by name")
